notebook/requirements/requirements.txt
amqp==5.0.6
anyjson==0.3.3
appdirs==1.4.4
billiard==3.6.4.0
black==21.5b1
cached-property==1.5.2
celery==5.1.0
certifi==2020.12.5
cffi==1.14.5
cfgv==3.3.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.7
distlib==0.3.1
Django==2.2.20
django-debug-toolbar==3.2.1
django-filter==2.4.0
djangorestframework==3.12.4
drf-yasg==1.20.0
Faker==8.2.0
filelock==3.0.12
flake8==3.9.2
gevent==21.1.2
greenlet==1.1.0
gunicorn==20.1.0
identify==2.2.4
idna==2.10
importlib-metadata==4.0.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jwt==1.2.0
kombu==5.1.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==20.9
pathspec==0.8.1
pre-commit==2.12.1
prompt-toolkit==3.0.18
psycopg2==2.7.7
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.4.4
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
virtualenv==20.4.6
wcwidth==0.2.5
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
